Vivian MEAZZA: a. A new attribute 'supercharger' has been added to the aircraft config file. When this is true, the wastegate no longer controls the supercharger output.
b. A Boost Control has been implemented in Nasal. When a preset boost value
is exceeded, the Boost Control acts to reduce throttle opening. This action
can be overridden by the Boost Control Cutout which allows maximum boost.

(requires a change to yasim that has yet to be discussed/reviewed)